Add StdHeaders for standard containers usage. Cleanup massive usage of Vector, String etc. Fix issue with Win32 window trimming the Title to only 1 character. Add Size and SizeF support under Math.

+++ Source/Alimer/Renderer/Renderer.h
@@ -47,13 +47,13 @@ namespace Alimer
 	};
 
 	/// Parameter indices in constant buffers used by high-level rendering.
-	static const uint32_t VS_FRAME_VIEW_MATRIX = 0;
+	static const size_t VS_FRAME_VIEW_MATRIX = 0;
 	static const uint32_t VS_FRAME_PROJECTION_MATRIX = 1;
 	static const uint32_t VS_FRAME_VIEWPROJ_MATRIX = 2;
 	static const uint32_t VS_FRAME_DEPTH_PARAMETERS = 3;
-	static const uint32_t VS_OBJECT_WORLD_MATRIX = 0;
+	static const size_t VS_OBJECT_WORLD_MATRIX = 0;
 	static const uint32_t VS_LIGHT_SHADOW_MATRICES = 0;
-	static const uint32_t PS_FRAME_AMBIENT_COLOR = 0;
+	static const size_t PS_FRAME_AMBIENT_COLOR = 0;
 	static const uint32_t PS_LIGHT_POSITIONS = 0;
 	static const uint32_t PS_LIGHT_DIRECTIONS = 1;
 	static const uint32_t PS_LIGHT_ATTENUATIONS = 2;
@@ -100,7 +100,7 @@ namespace Alimer
 		~Renderer();
 
 		/// Set number, size and format of shadow maps. These will be divided among the lights that need to render shadow maps.
-		void SetupShadowMaps(size_t num, int size, ImageFormat format);
+		void SetupShadowMaps(uint32_t num, uint32_t size, ImageFormat format);
 		/// Prepare a view for rendering. Convenience function that calls CollectObjects(), CollectLightInteractions() and CollectBatches() in one go. Return true on success.
 		bool PrepareView(Scene* scene, Camera* camera, const std::vector<PassDesc>& passes);
 		/// Initialize rendering of a new view and collect visible objects from the camera's point of view. Return true on success (scene, camera and octree are non-null.)
